+++ src/detections/print.rs
@@ -331,7 +331,7 @@ impl AlertMessage {
     /// ERRORメッセージを表示する関数
     pub fn alert(contents: &str) -> io::Result<()> {
         write_color_buffer(
-            BufferWriter::stderr(ColorChoice::Always),
+            &BufferWriter::stderr(ColorChoice::Always),
             None,
             &format!("[ERROR] {}", contents),
         )
@@ -340,7 +340,7 @@ impl AlertMessage {
     /// WARNメッセージを表示する関数
     pub fn warn(contents: &str) -> io::Result<()> {
         write_color_buffer(
-            BufferWriter::stderr(ColorChoice::Always),
+            &BufferWriter::stderr(ColorChoice::Always),
             None,
             &format!("[WARN] {}", contents),
         )