+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Updated the default usage and help menu. (#387) (@hitenkoku)
 - Added default details output based on `rules/config/default_details.txt` when no `details` field in a rule is specified. (i.e. Sigma rules) (#359) (@hitenkoku)
 - Added saved file size output when `output` is specified. (#595) (@hitenkoku)
-- Ignored loading yml file in `rules/tools/sigmac/testfiles` (#602) (@hitenkoku)
+- Ignored loading yml file in `rules/tools/sigmac/testfiles`. (#602) (@hitenkoku)
 
 **Bug Fixes:**